Revert "Merge branch 'api/v1/instance' into fix-spread-and-rest-based-binding"
This reverts commit 925eef2d5c02ab271b65054f1471309d2b530997, reversing
changes made to 85b0ac44a7a8079661943620985a76ec399a5a2c.
